Added Configs, Fully Wired Groups Service, Implemented Directory Service
This commit put all the Groups Service Components into play. First wiring a directory which is an interface to all of the actual google components, then a groups service, which will do the proper work eventually interfacing between google and the database(currently not implemented), next will be the simple(Just Google) members and users interface with lots more testing, as this is woefully under documented.

The test routes serve to show that the service is running correctly , and all other actions will be idempotent in nature. Finally I need to make sure to decouple and make the code as resilient as possible so that it can handle peak workflow.