Testing for Cors, Json, and GroupsServiceRoutes
Full routing is complete for the groups module atleast in testing but error handling need to integrated in some way. We need to containerize errors into the independent modules so that their is no leakage, and Futures allow for leakage, as the Failure could be an error we want contained to that module.

Responses are consistent and each piece is replaceable. Leaving lots of room for internal builds.