Implemented MembersDirectory, MemberTypes, And Making Movement Between Modules Smooth
The routes have been made so that they are all handled, MemberRoles and MemberTypes were introduced to allow us to deal with Google utilizing Strings rather than some type system. So we now are using enumeratum to keep all values acceptable. Email parsing was baked into the modules as necessary to make sure information was passed as email and not just as the email prefix.