Added Full Functional Error Handling
By utilizing a set of known abstractions that directly relate to responses we are able to build a fully functional system that splits errors into their own returnable subsystem that is handled seperately from business logic. Therefore allowing the service to thread the disjunction to a route through independent control flow systems.